requirements.txt
torch==1.0.0
torchvision==0.2.1
runway-python
dominate
pyssim
visdom